FairGig/services/earnings-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.5.1
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
